Dinesh

+++ components/Board.py
@@ -1,9 +1,10 @@
 import numpy as np
 
+
 class Board:
     def __init__(self, board=None, board_size=4) -> None:
         self.board_size = board.board_size if board else board_size
-        self.values = board.values.copy() if board else np.zeros((self.board_size, self.board_size))
+        self.values = board.values.copy() if board else np.zeros((self.board_size, self.board_size), dtype=int)
         self.score = 0
 
     def __repr__(self) -> str:
@@ -40,18 +41,18 @@ class Board:
         """compress all the numbers on one side of the board.
         """
         new_row = [i for i in row if i != 0]
-        return new_row + [0] * (self.board_size-len(new_row))
+        return new_row + [0] * (self.board_size - len(new_row))
 
     def _swiperow(self, row, score_update=False) -> list:
         """return the row after a swipe from right to left.
         """
         row = self._compress(row)
-        for i in range(len(row)-1):
-            if row[i] == row[i+1]:
-                row[i] = row[i]*2
+        for i in range(len(row) - 1):
+            if row[i] == row[i + 1]:
+                row[i] = row[i] * 2
                 if score_update:
                     self.score += row[i]
-                row[i+1] = 0
+                row[i + 1] = 0
         return self._compress(row)
 
     def _swiperowEfficient(self, row, score_update):
@@ -64,15 +65,13 @@ class Board:
         padded_array[:new_row.shape[0]] = new_row
         return padded_array
 
-
     def _swipeLeft(self, board_values, score_update) -> np.ndarray:
         new_board = np.zeros_like(board_values)
         for i in range(self.board_size):
             new_board[i, :] = self._swiperow(board_values[i, :], score_update)
         return new_board
-    
-        
-    def Swipe(self, direction, score_update, inplace=True): 
+
+    def Swipe(self, direction, score_update, inplace=True):
         moves2rot = {'left': (0, 4),
                      'up': (1, 3),
                      'right': (2, 2),
